% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullEntity.R
\name{getFullEntity}
\alias{getFullEntity}
\title{Retrieve full entity data}
\usage{
getFullEntity(link, format = "json", yourAPIKey = yourAPIKey)
}
\arguments{
\item{link}{link to the BioOntology entity}

\item{format}{Choose which format to get the results in; default is JSON, or XML}

\item{yourAPIKey}{manually include you API key for BioOntology here}
}
\description{
Get class and properties of an entity, using the BioOntology Class URI.
The results are also parsed from json or xml to provide readable results.
}
\details{
Before using this function, be sure to add your BioOntology API key
using the yourAPIKey variable. See ?yourAPIKey for more instructions.
}
\examples{
\dontrun{
# getFullEntity("http://data.bioontology.org/ontologies/RCD/classes/http\%3A\%2F\%2Fpurl.bioontology.org\%2Fontology\%2FRCD\%2FX78Tv")
}
}
