% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listUnique.R
\name{listUnique}
\alias{listUnique}
\title{Get unique list terms}
\usage{
listUnique(list)
}
\arguments{
\item{list}{list containing lists which are to be used}
}
\description{
Convenience function for converting list of lists into an unique only list.
}
