% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontTermAgg.R
\name{ontTermAgg}
\alias{ontTermAgg}
\title{Get all results of term}
\usage{
ontTermAgg(searchTerm, yourAPIKey, format = "json")
}
\arguments{
\item{searchTerm}{Enter the term you want the combined data for}

\item{yourAPIKey}{manually include you API key for BioOntology here}

\item{format}{Choose which format to get the results in; default is JSON, or XML}
}
\description{
Higher order function to essentially run the same code from the vignette as a function
}
