% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yourAPIKey.R
\name{yourAPIKey}
\alias{yourAPIKey}
\title{API Key for BioOntology}
\usage{
yourAPIKey(apikey = "ENTER_API_KEY_HERE")
}
\arguments{
\item{apikey}{Enter your API key}
}
\description{
Enter your personal key to allow access via additional functions
}
\details{
To get an API key, create an account on https://bioportal.bioontology.org/login
Go to 'Account Settings' via your username, there you will find the API Key
}
